added serve and staff

+++ bin/puppy.js
@@ -1,49 +1,69 @@
 #! /usr/bin/env node
 
+const fs = require('fs')
 const path = require('path')
-const spawn = require('child_process').spawn
 const chalk = require('chalk')
-const figlet = require('figlet')
+const spawn = require('child_process').spawn
+const minimist = require('minimist')
+const findFreePort = require('find-free-port')
 
-function getPuppy () {
-  return process.env.HEADLESS === 'true'
-    ? `
-                       \\
-         >--,-'\`-'*_*'\`\`---.
-         |\\_* _*'-'         '
-        /    \`               \\
-        \\.         _ .       /
-         '\`._     /   )     /
-          \\  |\`-,-|  /c-'7 /
-           ) \\ (_,| |   / (_
-          ((_/   ((_;)  \\_)))`
-
-    : `          ,_____ ,
-         ,._ ,_. 7\\
-        j \`-'     /
-        |o_, o    \\
-       .\`_y_\`-,'   !
-       |/   \`, \`._ \`-,
-       |_     \\   _.'*\\
-         >--,-'\`-'*_*'\`\`---.
-         |\\_* _*'-'         '
-        /    \`               \\
-        \\.         _ .       /
-         '\`._     /   )     /
-          \\  |\`-,-|  /c-'7 /
-           ) \\ (_,| |   / (_
-          ((_/   ((_;)  \\_)))`
-}
+const logo = require('../src/logo')
+const help = require('../src/help')
+const version = require('../package').version
+const puppyConfig = require('../puppy.config.js')
 
 ;(async () => {
-  console.log(chalk.cyan(getPuppy()))
-  figlet('puppy.js', (_, data) => console.log(chalk.red(data)))
-
-  const config = path.resolve(__dirname, '..', 'config/jest.config.js')
+  // files
   const serverFile = path.resolve(__dirname, '..', 'src/server.js')
+  const jestConfigFile = path.resolve(__dirname, '..', 'config/jest.config.js')
 
-  const server = spawn(`node`, ['--inspect', serverFile], {pwd: process.cwd(), stdio: 'inherit'})
-  const jest = spawn('jest', ['-i', '-c', config, '--rootDir', process.cwd(), ...process.argv.slice(2)], {stdio: 'inherit'})
+  // arguments
+  const arguments = minimist(process.argv.slice(2), {boolean: ['h', 'help', 'version', 'headless']})
+
+  if (arguments.version) {
+    return console.log('Version:', version)
+  }
+
+  if (arguments.help) {
+    return console.log(help())
+  }
+
+  const configFile = path.resolve(process.cwd(), arguments.config || 'puppy.config.js')
+
+  if (fs.existsSync(configFile)) {
+    Object.assign(puppyConfig, require(configFile))
+  }
+
+  const WS = arguments['ws'] || puppyConfig['ws']
+  const API = arguments['api'] || puppyConfig['api']
+
+  const PORT = arguments['port'] || puppyConfig['port']
+  const WS_PORT = arguments['ws-port'] || arguments['api-port'] || arguments['port'] || puppyConfig['ws-api'] || puppyConfig['api-port'] || puppyConfig['port']
+  const API_PORT = arguments['api-port'] || arguments['port'] || puppyConfig['api-port'] || puppyConfig['port']
+  const INTERNAL_PORT = (await findFreePort(65000, 65535)).pop()
+
+  const VERBOSE = arguments['verbose'] || puppyConfig['verbose'] || false
+
+  const WS_URL = arguments['ws-url'] || puppyConfig['ws-url']
+  const INDEX_FILE = arguments['index-file'] || puppyConfig['index-file']
+  const STATIC_DIR = arguments['static-dir'] || puppyConfig['static-dir']
+
+  console.log(chalk.cyan(logo(arguments.headless)))
+
+  const server = spawn(`node`, ['--inspect', serverFile], {
+    pwd: process.cwd(),
+    stdio: 'inherit',
+    env: Object.assign({}, process.env, {WS, API, PORT, WS_PORT, API_PORT, INTERNAL_PORT, VERBOSE, WS_URL, INDEX_FILE, STATIC_DIR})
+  })
+
+  if (arguments._.includes('serve') || arguments._.includes('s')) {
+    return console.log('serving only')
+  }
+
+  const jest = spawn('jest', ['-i', '-c', jestConfigFile, '--rootDir', process.cwd()], {
+    stdio: 'inherit',
+    env: Object.assign({}, process.env, {WS, API, PORT, WS_PORT, API_PORT, INTERNAL_PORT, VERBOSE, WS_URL, INDEX_FILE, STATIC_DIR})
+  })
 
   process.on('SIGHUP', () => server.kill('SIGHUP'))
   process.on('SIGTERM', () => server.kill('SIGHUP'))
